refactor of index.ts
